icpc_prac23

+++ vjudge/ICPC_prac_23/D.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main(){
 
-    int n{0},m{0},t{0};
+    long long int n{0},m{0},t{0};
     cin>>n>>m;
     if(m>=n &&n>=1){
         t=n+(3*(n-1));
